+++ packages/jsonapi/src/Validation/Patterns.php
@@ -16,4 +16,14 @@ class Patterns
      * This is the intersection of valid PHP property names and valid JSON:API member names.
      */
     public const PROPERTY_NAME = '[a-z]\w*';
+
+    /**
+     * A comma separated list of one or more property names.
+     */
+    public const PROPERTY_NAME_LIST = self::PROPERTY_NAME.'(,'.self::PROPERTY_NAME.')*';
+
+    /**
+     * A dot separated list of one or more property names.
+     */
+    public const PROPERTY_PATH = self::PROPERTY_NAME.'(.'.self::PROPERTY_NAME.')*';
 }